+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMViewScope.swift
@@ -556,7 +556,16 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
             dependencies.fatalErrorContext.view = event
 
             // Track this view in Session Ended metric:
-            dependencies.sessionEndedMetric.track(view: event, in: self.context.sessionID)
+            dependencies.sessionEndedMetric.track(
+                view: event,
+                instrumentationType: (command as? RUMStartViewCommand)?.instrumentationType,
+                in: self.context.sessionID
+            )
+
+            // Update the state of the view in watchdog termination monitor
+            // if a watchdog termination occurs in this session, in the next session
+            // a watchdog termination event will be sent using saved view event.
+            dependencies.watchdogTermination?.update(viewEvent: event)
         } else { // if event was dropped by mapper
             version -= 1
         }