+++ DatadogRUM/Tests/SDKMetrics/SessionEndedMetricControllerTests.swift
@@ -19,18 +19,20 @@ class SessionEndedMetricControllerTests: XCTestCase {
 
         // Given
         let controller = SessionEndedMetricController(telemetry: telemetry)
-        controller.startMetric(sessionID: sessionID, precondition: .mockRandom(), context: .mockRandom())
+        controller.startMetric(sessionID: sessionID, precondition: .mockRandom(), context: .mockRandom(), tracksBackgroundEvents: .mockRandom())
 
         // When
-        viewIDs.forEach { controller.track(view: .mockRandomWith(sessionID: sessionID, viewID: $0), in: sessionID) }
+        viewIDs.forEach { controller.track(view: .mockRandomWith(sessionID: sessionID, viewID: $0), instrumentationType: nil, in: sessionID) }
         errorKinds.forEach { controller.track(sdkErrorKind: $0, in: sessionID) }
+        controller.track(missedEventType: .action, in: sessionID)
         controller.trackWasStopped(sessionID: sessionID)
-        controller.endMetric(sessionID: sessionID)
+        controller.endMetric(sessionID: sessionID, with: .mockRandom())
 
         // Then
         let metric = try XCTUnwrap(telemetry.messages.lastSessionEndedMetric)
         XCTAssertEqual(metric.viewsCount.total, viewIDs.count)
         XCTAssertEqual(metric.sdkErrorsCount.total, errorKinds.count)
+        XCTAssertEqual(metric.noViewEventsCount.actions, 1)
         XCTAssertEqual(metric.wasStopped, true)
     }
 
@@ -40,18 +42,18 @@ class SessionEndedMetricControllerTests: XCTestCase {
 
         // When
         let controller = SessionEndedMetricController(telemetry: telemetry)
-        controller.startMetric(sessionID: sessionID1, precondition: .mockRandom(), context: .mockRandom())
-        controller.startMetric(sessionID: sessionID2, precondition: .mockRandom(), context: .mockRandom())
+        controller.startMetric(sessionID: sessionID1, precondition: .mockRandom(), context: .mockRandom(), tracksBackgroundEvents: .mockRandom())
+        controller.startMetric(sessionID: sessionID2, precondition: .mockRandom(), context: .mockRandom(), tracksBackgroundEvents: .mockRandom())
         // Session 1:
-        controller.track(view: .mockRandomWith(sessionID: sessionID1), in: sessionID1)
+        controller.track(view: .mockRandomWith(sessionID: sessionID1), instrumentationType: nil, in: sessionID1)
         controller.track(sdkErrorKind: "error.kind1", in: sessionID1)
         controller.trackWasStopped(sessionID: sessionID1)
         // Session 2:
         controller.track(sdkErrorKind: "error.kind2", in: sessionID2)
         // Send 1st and 2nd:
-        controller.endMetric(sessionID: sessionID1)
+        controller.endMetric(sessionID: sessionID1, with: .mockRandom())
         let metric1 = try XCTUnwrap(telemetry.messages.lastSessionEndedMetric)
-        controller.endMetric(sessionID: sessionID2)
+        controller.endMetric(sessionID: sessionID2, with: .mockRandom())
         let metric2 = try XCTUnwrap(telemetry.messages.lastSessionEndedMetric)
 
         // Then
@@ -72,19 +74,21 @@ class SessionEndedMetricControllerTests: XCTestCase {
 
         // When
         let controller = SessionEndedMetricController(telemetry: telemetry)
-        controller.startMetric(sessionID: sessionID1, precondition: .mockRandom(), context: .mockRandom())
-        controller.startMetric(sessionID: sessionID2, precondition: .mockRandom(), context: .mockRandom())
+        controller.startMetric(sessionID: sessionID1, precondition: .mockRandom(), context: .mockRandom(), tracksBackgroundEvents: .mockRandom())
+        controller.startMetric(sessionID: sessionID2, precondition: .mockRandom(), context: .mockRandom(), tracksBackgroundEvents: .mockRandom())
         // Track latest session (`sessionID: nil`)
-        controller.track(view: .mockRandomWith(sessionID: sessionID2), in: nil)
+        controller.track(view: .mockRandomWith(sessionID: sessionID2), instrumentationType: nil, in: nil)
         controller.track(sdkErrorKind: "error.kind1", in: nil)
+        controller.track(missedEventType: .resource, in: nil)
         controller.trackWasStopped(sessionID: nil)
         // Send 2nd:
-        controller.endMetric(sessionID: sessionID2)
+        controller.endMetric(sessionID: sessionID2, with: .mockRandom())
         let metric = try XCTUnwrap(telemetry.messages.lastSessionEndedMetric)
 
         // Then
         XCTAssertEqual(metric.viewsCount.total, 1)
         XCTAssertEqual(metric.sdkErrorsCount.total, 1)
+        XCTAssertEqual(metric.noViewEventsCount.resources, 1)
         XCTAssertEqual(metric.wasStopped, true)
     }
 
@@ -98,15 +102,17 @@ class SessionEndedMetricControllerTests: XCTestCase {
         callConcurrently(
             closures: [
                 { controller.startMetric(
-                    sessionID: sessionIDs.randomElement()!, precondition: .mockRandom(), context: .mockRandom()
+                    sessionID: sessionIDs.randomElement()!, precondition: .mockRandom(), context: .mockRandom(), tracksBackgroundEvents: .mockRandom()
                 ) },
-                { controller.track(view: .mockRandom(), in: sessionIDs.randomElement()!) },
+                { controller.track(view: .mockRandom(), instrumentationType: nil, in: sessionIDs.randomElement()!) },
                 { controller.track(sdkErrorKind: .mockRandom(), in: sessionIDs.randomElement()!) },
                 { controller.trackWasStopped(sessionID: sessionIDs.randomElement()!) },
-                { controller.track(view: .mockRandom(), in: nil) },
+                { controller.track(view: .mockRandom(), instrumentationType: nil, in: nil) },
                 { controller.track(sdkErrorKind: .mockRandom(), in: nil) },
+                { controller.track(missedEventType: .action, in: sessionIDs.randomElement()!) },
+                { controller.track(missedEventType: .resource, in: nil) },
                 { controller.trackWasStopped(sessionID: nil) },
-                { controller.endMetric(sessionID: sessionIDs.randomElement()!) },
+                { controller.endMetric(sessionID: sessionIDs.randomElement()!, with: .mockRandom()) },
             ],
             iterations: 100
         )