+++ DatadogSessionReplay/SRSnapshotTests/SRFixtures/Sources/SRFixtures/ViewControllers/ImagesViewControllers.swift
@@ -10,8 +10,6 @@ internal class ImagesViewController: UIViewController {
     @IBOutlet weak var customButton: UIButton!
     @IBOutlet weak var customImageView: UIImageView!
     @IBOutlet weak var contentImageView: UIImageView!
-    @IBOutlet weak var tabBar: UITabBar!
-    @IBOutlet weak var navigationBar: UINavigationBar!
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -21,10 +19,6 @@ internal class ImagesViewController: UIViewController {
 
         let color = UIColor(white: 0, alpha: 0.05)
         customButton.setBackgroundImage(UIImage(color: color), for: .normal)
-
-        tabBar.backgroundImage = UIImage(color: color)
-        tabBar.selectedItem = tabBar.items?.first
-        navigationBar.setBackgroundImage(UIImage(color: color), for: .default)
 
         let image = UIImage(named: "dd_logo", in: .module, with: nil)
         customImageView.image = image?.withRenderingMode(.alwaysTemplate)