+++ DatadogSessionReplay/Sources/Feature/SessionReplayFeature.swift
@@ -39,7 +39,6 @@ internal class SessionReplayFeature: SessionReplayConfiguration, DatadogRemoteFe
         let recorder = try Recorder(
             snapshotProcessor: snapshotProcessor,
             resourceProcessor: resourceProcessor,
-            telemetry: core.telemetry,
             additionalNodeRecorders: configuration._additionalNodeRecorders
         )
         let scheduler = MainThreadScheduler(interval: 0.1)
@@ -59,7 +58,8 @@ internal class SessionReplayFeature: SessionReplayConfiguration, DatadogRemoteFe
             rumContextObserver: contextReceiver,
             srContextPublisher: SRContextPublisher(core: core),
             recorder: recorder,
-            sampler: Sampler(samplingRate: configuration.debugSDK ? 100 : configuration.replaySampleRate)
+            sampler: Sampler(samplingRate: configuration.debugSDK ? 100 : configuration.replaySampleRate),
+            telemetry: core.telemetry
         )
         self.requestBuilder = SegmentRequestBuilder(
             customUploadURL: configuration.customEndpoint,